Rename receipt
Issue #224
